Fix bug in keywords tab when the same keyword is in multiple books.
The keywords tab did not limit the search of blocks to the current book
when selecting a keyword.
When having multiple books with the same keyword this could produce an
exception within the internal chapters cache when selecting this keyword.

The solution is to properly limit the scope of the catalog query to the
current book.