ftw.publisher support: make sure the layout adapters are respected.
IDataCollector for supporting publishing books (ftw.publisher) with
custom layouts and schema extender fields bound on layout request layer.

When a custom layout is activated in the book, the request provides the
layout interface as soon as traversing the book.
The layout may add schema extender fields when the request provides the
layout interface.
In order to publish those field values correctly (with the standard
FieldData collector), we need to provide the request layer before
extracting data (sender) or setting data (receiver).

Schema extender fields may be applied to not only the book itself but also
to any content within the book. Therfore this collector should be used
for all objects.

The collector may be triggered when the book is traversed but also when
it wasn't, therefore we cannot rely on the default book traversal adapter
applying the layer interfaces to the request.