fix listingblock tests
Somehow something changed regarding newlines in lxml, leading to
failing tests:

It seems as if newlines were cleaned up before and now longer are.
The problem seems to depend on the lxml version and therefore not appear
on every installation.

1) `parsed_html`-decorator: The decorator parses HTML to an lxml
document, wich is then modified by the wrapped function and the
decorator then renders the document. For that the HTML is wrapped with a
`<div>` in order to avoid a multiple-root-node error. When removing the
`<div>` afterwards, we need to make sure that we do that in a multiline
supporting way (`re.DOTALL`) in order to support multilines in the
HTML. I'm not sure why this was not a problem before.

2) Since something in lxml changed we now have line breaks in the
resulting HTML. We just remove them in order to support both lxml
versions - we do not really care.