PDF: use external links for non-PDF content.
When a content is linked in a book and the target content is included
in the same PDF, an PDF-internal reference is used.
For external resources a URL-based link is used.

When linking files, which are stored in a listing block in the book,
we still want to use an external URL-based link because files are not
embedded in the PDF, so that file content is not part of the PDF.